added button functionality to ui_elements and a button to welcome window
