+++ src/modules/ui_elements/ui_elements.h
@@ -1,0 +1,6 @@
+#ifndef UI_ELEMENTS_H
+#define UI_ELEMENTS_H
+
+#include "modules/ui_elements/button.h"
+
+#endif // UI_ELEMENTS_H
